Trim examples to one script per workflow
Keep dataset, dataloader, transform, and TorchVision integration as the
example scripts, and drop the corpus-checkout scripts together with the
Font Awesome, Material Design Icons, and Source Han Code JP submodules.
Google Fonts remains the only data submodule.

Remove the documentation example gallery as well. It described this
repository's own scripts rather than the library, and the guide already
covers the same workflows.
