+++ .ignore
@@ -1,9 +1,6 @@
-# Large optional data submodules. Keep them out of default repository searches;
-# search them explicitly with `rg -u` or a direct path when needed.
-data/adobe/source-han-code-jp/
-data/fortawesome/font-awesome/
+# Large optional data submodule. Keep it out of default repository searches;
+# search it explicitly with `rg -u` or a direct path when needed.
 data/google/fonts/
-data/google/material_design_icons/
 
 # Keep shared repository tooling discoverable in default ripgrep file searches.
 !/.devcontainer/
