+++ examples/transform.py
@@ -1,0 +1,51 @@
+from torch.utils.data import DataLoader
+from tqdm import tqdm
+
+from torchfont import GlyphData, Outline, pad_outlines
+from torchfont import transforms as T
+from torchfont.datasets import GlyphDataset
+
+
+def collate_fn(batch: list[GlyphData[Outline]]) -> Outline:
+    return pad_outlines([data.data for data in batch])
+
+
+def main() -> None:
+    transform = T.Compose(
+        [
+            T.LoadGlyph(location="random"),
+            T.RemoveOverlaps(),
+            T.QuadToCubic(merge_curves=True),
+        ]
+    )
+
+    dataset = GlyphDataset(
+        root="data/google/fonts",
+        patterns=(
+            "apache/*/*.ttf",
+            "ofl/*/*.ttf",
+            "ufl/*/*.ttf",
+            "!ofl/adobeblank/*.ttf",
+        ),
+        transform=transform,
+    )
+
+    dataloader = DataLoader(
+        dataset,
+        batch_size=64,
+        shuffle=True,
+        num_workers=8,
+        prefetch_factor=2,
+        collate_fn=collate_fn,
+    )
+
+    print(f"{len(dataset)=}")
+    print(f"{len(dataset.font_classes)=}")
+    print(f"{len(dataset.character_classes)=}")
+
+    for batch in tqdm(dataloader, desc="Iterating over datasets"):
+        _ = batch
+
+
+if __name__ == "__main__":
+    main()
